refactor: rename SocialButton-styling file
